Move print_param_types() to libapps, and give it indent argument
