+++ apps/build.info
@@ -31,7 +31,7 @@ $OPENSSLSRC={-
    join(' ', @opensslsrc); -}
 # Source for libapps
 $LIBAPPSSRC=apps.c apps_ui.c opt.c fmt.c s_cb.c s_socket.c app_rand.c \
-        bf_prefix.c columns.c
+        bf_prefix.c columns.c lib/app_params.c
 
 IF[{- !$disabled{apps} -}]
   LIBS{noinst}=libapps.a
